+++ packages/s2-react/src/hooks/useSpreadSheet.ts
@@ -5,6 +5,7 @@ import {
   SpreadSheet,
   TableSheet,
 } from '@antv/s2';
+import { useUpdate } from 'ahooks';
 import React from 'react';
 import type { BaseSheetComponentProps, SheetType } from '../components';
 import { getSheetComponentOptions } from '../utils';
@@ -23,16 +24,18 @@ export function useSpreadSheet(
   props: BaseSheetComponentProps,
   config: UseSpreadSheetConfig,
 ) {
+  const forceUpdate = useUpdate();
   const s2Ref = React.useRef<SpreadSheet>();
   const containerRef = React.useRef<HTMLDivElement>();
 
   const { spreadsheet: customSpreadSheet, dataCfg, options, themeCfg } = props;
   const { loading, setLoading } = useLoading(s2Ref.current, props.loading);
-  const { registerEvent } = useEvents(props);
   const pagination = usePagination(s2Ref.current, props);
   const prevDataCfg = usePrevious(dataCfg);
   const prevOptions = usePrevious(options);
   const prevThemeCfg = usePrevious(themeCfg);
+
+  useEvents(props, s2Ref.current);
 
   const renderSpreadSheet = React.useCallback(
     (container: HTMLDivElement) => {
@@ -53,11 +56,15 @@ export function useSpreadSheet(
     setLoading(true);
     s2Ref.current = renderSpreadSheet(containerRef.current);
     s2Ref.current.setThemeCfg(props.themeCfg);
-    registerEvent(s2Ref.current);
     s2Ref.current.render();
     setLoading(false);
+
+    // 子 hooks 内使用了 s2Ref.current 作为 dep
+    // forceUpdate 一下保证子 hooks 能 rerender
+    forceUpdate();
+
     props.getSpreadSheet?.(s2Ref.current);
-  }, [props, registerEvent, renderSpreadSheet, setLoading]);
+  }, [props, renderSpreadSheet, setLoading, forceUpdate]);
 
   // init
   React.useEffect(() => {