+++ s2-site/examples/case/data-preview/demo/index.tsx
@@ -556,7 +556,7 @@ const SortPopover = ({ fieldName, spreadsheet, modalCallbackRef }) => {
 
   modalCallbackRef.current = () => {
     if (changed.sort) {
-      spreadsheet.emit(S2Event.RANGE_SORT, {
+      spreadsheet.emit(S2Event.SORT_GROUP_SORT, {
         sortKey: fieldName,
         sortMethod: sort,
         sortBy: (obj) => {
@@ -569,7 +569,7 @@ const SortPopover = ({ fieldName, spreadsheet, modalCallbackRef }) => {
       });
     }
     if (changed.filter) {
-      spreadsheet.emit(S2Event.RANGE_FILTER, {
+      spreadsheet.emit(S2Event.FILTER, {
         filterKey: fieldName,
         // 将Object还原成数组
         filteredValues: Object.entries(filtered)